+++ Sources/ColorTokensKit/Platform/UIKit/UIColor+Dynamic.swift
@@ -1,0 +1,27 @@
+//
+//  UIColor+Dynamic.swift
+//  ColorTokensKit
+//
+
+#if canImport(UIKit)
+import UIKit
+
+public extension UIColor {
+    /// Initialize with light/dark mode colors
+    convenience init(
+        light lightModeColor: @escaping @autoclosure () -> UIColor,
+        dark darkModeColor: @escaping @autoclosure () -> UIColor
+    ) {
+        self.init { traitCollection in
+            switch traitCollection.userInterfaceStyle {
+            case .light:
+                return lightModeColor()
+            case .dark:
+                return darkModeColor()
+            default:
+                return lightModeColor()
+            }
+        }
+    }
+}
+#endif
